debt - cleanup file model a bit
